services/data-sync/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery[sqs]==5.3.6
kombu[sqs]==5.3.3
boto3==1.34.0
redis==5.0.1
psycopg2-binary==2.9.9
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
